+++ aiu_fms_testing_utils/testing/dpp/run_drive_paged_programs.py
@@ -9,15 +9,17 @@ from aiu_fms_testing_utils.testing.dpp.prepare_data import (
     prepare_test_prompts,
     resolve_dataset_path,
 )
+from aiu_fms_testing_utils.testing.dpp.prepare_model import (
+    _get_model_kwargs,
+    load_model,
+)
 from aiu_fms_testing_utils.testing.dpp.program_models import EnvConfig
 from aiu_fms_testing_utils.testing.dpp.sample_prompts import get_sampler
-from aiu_fms_testing_utils.utils import stagger_region, warmup_model
+from aiu_fms_testing_utils.utils import warmup_model
 from aiu_fms_testing_utils.utils.aiu_setup import aiu_dist_setup, dprint, local_rank
 from aiu_fms_testing_utils.utils.dpp_config import DPPRunnerConfig
-from aiu_fms_testing_utils.utils.env_utils import scoped_environ
 from aiu_fms_testing_utils.testing.dpp.constants import PAD_MULTIPLE
 
-from fms.models import get_model
 from fms.utils.generation import pad_input_ids
 
 import torch
@@ -25,7 +27,7 @@ from torch import distributed as dist
 from transformers import AutoTokenizer
 
 
-from typing import Any, Dict, List, Literal
+from typing import Any, Dict, List
 
 
 DEFAULT_CE_THRESHOLD = 2.5
@@ -63,111 +65,6 @@ def _get_distributed_kwargs(dist_timeout: str) -> Dict[str, Any]:
         "distributed_strategy": "tp",
         "group": dist.group.WORLD,
     }
-
-
-def _get_model_kwargs(model_variant: str) -> Dict[str, Any]:
-    """Constructs model loading kwargs based on whether variant is a path or ID.
-
-    Determines if the model_variant is a local filesystem path or a HuggingFace
-    model identifier, and returns the appropriate keyword arguments for model loading.
-
-    Args:
-        model_variant: Either a local path to model files or a HuggingFace model ID.
-
-    Returns:
-        Dictionary with either "model_path" (for local paths) or "variant"
-        (for HuggingFace IDs) as the key."""
-
-    model_kwargs = {}
-    if os.path.exists(model_variant):
-        model_kwargs["model_path"] = model_variant
-    else:
-        model_kwargs["variant"] = model_variant
-
-    return model_kwargs
-
-
-def _prepare_fp8_weights(model: torch.nn.Module) -> None:
-    """Converts model weights from bfloat16 to float16 for FP8 attention.
-
-    When using FP8 attention variants, this function converts all bfloat16 parameters
-    to float16. Issues a warning if any parameter values exceed the float16 range,
-    which may cause accuracy loss.
-
-    Args:
-        model: PyTorch model whose weights may need conversion."""
-
-    for name, param in model.named_parameters():
-        if param.dtype == torch.bfloat16:
-            if param.max() > torch.finfo(torch.float16).max:
-                dprint(
-                    f"[WARNING] You are casting param {name} to fp16, which will cause loss of accuracy."
-                )
-            param.data = param.data.to(dtype=torch.float16)
-
-
-def load_model(
-    device_type: Literal["cpu", "spyre"],
-    is_fp8: bool,
-    model_kwargs: Dict[str, Any],
-    distributed_kwargs: Dict[str, Any],
-    stagger_load: int,
-    model_config: DPPRunnerConfig,
-):
-    """Loads and optionally compiles a model for inference or validation.
-
-    Loads a model with the specified configuration. For Spyre/AIU models,
-    compiles the model using the sendnn backend with dynamic compilation enabled.
-    The scoped_environ context manager temporarily sets environment variables
-    from model_config during compilation to configure the compiler's behavior (e.g.,
-    program criteria, batch sizes, context lengths).
-
-    Args:
-        device_type: Target device for model execution. Options:
-            - "cpu": Load on CPU for validation (fp32, no compilation)
-            - "spyre": Load on CPU, compile for Spyre/AIU execution (fp16, with sendnn compilation)
-        is_fp8: If True, uses FP8 quantization (dtype=None for auto-detection).
-        model_kwargs: Dictionary with model loading parameters (variant or path).
-        distributed_kwargs: Dictionary with distributed training configuration.
-        stagger_load: Number of concurrent processes allowed during loading (0=unlimited).
-        model_config: DPPRunnerConfig instance with environment variable updates.
-
-    Returns:
-        torch.nn.Module: Loaded model in evaluation mode. Spyre models are compiled
-        with sendnn backend and may have FP8 weight conversion applied."""
-
-    if device_type not in ["cpu", "spyre"]:
-        raise ValueError(
-            f"device_type must be 'cpu' or 'spyre' for DPP, got '{device_type}'"
-        )
-
-    dtype = torch.float32 if device_type == "cpu" else torch.float16
-    if is_fp8:
-        dtype = None  # Let the model loading logic decide the appropriate FP8 dtype
-
-    with stagger_region(stagger_load):
-        model = get_model(
-            architecture="hf_pretrained",
-            device_type="cpu",
-            data_type=dtype,
-            fused_weights=False,
-            **model_kwargs,
-            **distributed_kwargs,
-        )
-
-    model.eval()
-
-    if device_type != "spyre":
-        return model
-
-    with scoped_environ(model_config.env_updates()):
-        # Temporarily set environment variables needed for compile
-        model.compile(backend="sendnn", options={"sendnn.dynamic": True})
-
-    if is_fp8:
-        _prepare_fp8_weights(model)
-
-    return model
 
 
 def setup_environment(
@@ -266,7 +163,7 @@ def run_dpp(
     )
 
     # Model Loading
-    model_kwargs = _get_model_kwargs(model_variant=model_variant)
+    model_kwargs = _get_model_kwargs(model_variant)
     distributed_kwargs = _get_distributed_kwargs(dist_timeout) if distributed else {}
 
     # Setup model config
